requirements.txt
requests
chromadb
streamlit
groq
python-dotenv